fix: Tag 테이블 이름 변경
