--- src/main/java/com/example/humorie/consultant/review/entity/ReviewTag.java
@@ -1,33 +1,0 @@
-package com.example.humorie.consultant.review.entity;
-
-import com.example.humorie.account.entity.AccountDetail;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.Set;
-
-@Entity
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ReviewTag {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private String tagName;
-
-    private String tagContent;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private AccountDetail account;
-
-    @ManyToMany(fetch = FetchType.LAZY)
-    private Set<Review> reviews;
-
-}
