+++ src/main/java/com/example/humorie/consultant/review/mapper/TagMapper.java
@@ -2,7 +2,7 @@ package com.example.humorie.consultant.review.mapper;
 
 import com.example.humorie.consultant.review.dto.TagReq;
 import com.example.humorie.consultant.review.dto.TagRes;
-import com.example.humorie.consultant.review.entity.ReviewTag;
+import com.example.humorie.consultant.review.entity.Tag;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
@@ -15,12 +15,12 @@ public interface TagMapper {
     TagMapper INSTANCE = Mappers.getMapper(TagMapper.class);
 
     @Mapping(source = "id", target = "tagId")
-    TagRes toTagRes(ReviewTag tag);
+    TagRes toTagRes(Tag tag);
 
     @Mapping(target = "id", ignore = true)
     @Mapping(target = "reviews", ignore = true)
-    ReviewTag toReviewTag(TagReq tagReq);
+    Tag toReviewTag(TagReq tagReq);
 
-    List<TagRes> toTagResList(List<ReviewTag> tags);
+    List<TagRes> toTagResList(List<Tag> tags);
 
 }
