+++ src/main/java/com/example/humorie/consultant/review/repository/TagRepository.java
@@ -1,7 +1,7 @@
 package com.example.humorie.consultant.review.repository;
 
 import com.example.humorie.account.entity.AccountDetail;
-import com.example.humorie.consultant.review.entity.ReviewTag;
+import com.example.humorie.consultant.review.entity.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,13 +9,13 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface TagRepository extends JpaRepository<ReviewTag, Long> {
+public interface TagRepository extends JpaRepository<Tag, Long> {
 
-    Optional<ReviewTag> findByTagNameAndAccount(String tagName, AccountDetail accountDetail);
+    Optional<Tag> findByTagNameAndAccount(String tagName, AccountDetail accountDetail);
 
-    List<ReviewTag> findByAccount(AccountDetail account);
+    List<Tag> findByAccount(AccountDetail account);
 
-    Optional<ReviewTag> findByIdAndAccount(long tagId, AccountDetail account);
+    Optional<Tag> findByIdAndAccount(long tagId, AccountDetail account);
 
     boolean existsByTagNameAndAccount(String tagName, AccountDetail account);
